Ian Jeffray:
I've been prattling with libscl yet again and found the root cause of the
backtrace problem.  The language descriptor block was wrong because
GCC creates areas called C$$CodeX where X is a number from 1 upward.
This invariably meant that at link time C$$Code$$Limit was
useless. The backtracer in SharedCLib appears to ignore function calls
outside the language block, so ignored the whole of the code built
with GCC. ;-)


I've bodged this by changing to use the Image$$RO$$Limit block, which
isn't superb, but is the next best fix.  (We can't just change to use
C$$Code1$$Limit as that's still pandering to the oddness of gcc).


I've also cleaned up language entry and the vector table, so now we
get a clean traceback to the language initialisation which I've called
___init now.  (Why? Just because I personally think it's good to see
an instantly recognisable symbol which should appear at the end of all
backtraces showing that we've not farted on our stack at any point).