Upgrade to CMunge v0.67 which (apart from a few changes, see README file) contains now all the GCCSDK specific support.  Apart from that, several bug fixes and feature improvements were implemented.  See Docs/History for all the details.
Note we are now using libscl as runtime library and that the toolchain
default (which got changed in the GCCSDK 3.4.4 release 3 into 'gcc') is
now back to 'norcroft' in order to have consistency between the GCCSDK
version of CMunge and the official CMunge.

Thanks to Justin Fletcher for his support in all this.