-Bug 161 fix.
-The bug manifests itself when a function with more than 4 arguments is
optimised by GCC so that no stack frame is generated. Such a function
is getopt_long (libunixlib/unix/getop1.c) which has 6 arguments. Normally,
the last 2 arguments are accessed via fp(r11), but due to the lack of a
stack frame, they are instead accessed as an offset to sp(r13). To
workout this offset, the registers that are stored on the stack during
function entry must be taken into account, however, the PIC register was
not being flagged as used early enough for GCC to take it into account
thereby producing an incorrect offset.

-Change parenthesis nesting to remove compiler warning.