Merge r2212/r2720/r2971 from branches/release_3_4_6 to trunk:
This involved splitting riscos.c into throwback specific code (as riscos.c)
and into initialisation code + riscos_convert_filename (riscos-gcc.c).
Created in libcpp a new callback 'throwback' which targets can use to
listen to the info/error/warning messages given (this differs from 'error'
callback as it does not intercept those messages). In frontend the
'throwback' callback is filled in when TARGET_ERROR_THROWBACK is defined
by the target and that also specifies the routine which needs to be called.