The compiler is very specific about what instructions it expects to see in a function prologue when emitting ARM unwind directives. A number of instructions are generated to support stack extension that the compiler doesn't expect to see:
* gcc/config/arm/arm.c(arm_unwind_emit): The rt_stkovf pattern is
defined as a PARALLEL - ignore these. If, in the future, the
compiler generates a valid PARALLEL instruction that requires an
unwind directive, then we'll have to be more specific.
* (arm_unwind_emit_set): Ignore any instructions in the prologue
that are related to stack extension.
Ignore any "mov r9, sp" instructions as these cause extra ".movsp"
unwind directives to be emitted. There will be a ".movsp ip"
instruction at the top of the function. Any instances of ".movsp r9"
in the body of the function generate an error from the assembler -
you can't change the register once it's been set. 
The last one doesn't seem specific to RISC OS - I'm not sure why it
isn't a problem elsewhere.