Merge r5277 from trunk/gcc4 to branches/developers/joty/gcc47.
------------------------------------------------------------------------
r5277 | lee | 2011-08-11 20:58:42 +0200 (Thu, 11 Aug 2011) | 20 lines

It's possible when lazy binding for the dynamic resolver to require a stack
extension. Previously, I assumed that as far as the dynamic linker is
concerned a stack extension would only occur when libdl was used, however,
running GCC shows that this is not the case. Although there were no errors
or side effects, it can't be a good thing to ignore a request to extend
the stack.

* d-link/libdl/dlib.c(_dlopen): Instead of switching to the UnixLib versions
  of __rt_stkovf_split_small, __rt_stkovf_split_big and malloc here...
* d-link/boot1.c(_dl_call_ctors): ... do it here instead.
  (_dl_malloc_function): Change visibility to hidden.
  (_dl_stkovf_split_small_function): Likewise.
  (_dl_stkovf_split_big_function): Likewise.
* d-link/riscos/sysdep.h(ATTRIBUTE_HIDDEN): Define shorthand version of
  hidden visibility attribute.

Now, instead of the UnixLib functions being available when dlopen is first
used (if ever), they will be switched in straight after UnixLib has
initialised.