Add Mono Port.
This includes RISC OS Dot Net, a library/assembly to access
the RISC OS API. It's far from complete and needs more work.
There are elements of a class library that I've used for
testing - these should probably be split off into a seperate
library eventually.

This is a Work In Progress, I intend to expand
RISC OS Dot Net and add more of the C# tools as I figure
out which ones work, what they do and if they're of any use
on RISC OS.

To build RISC OS Mono, you need a native C# compiler which
means installing Mono natively. Unfortunately, Mono 2.6.7
(included with Ubuntu 10.10 & 11.04) causes segmentation
faults, so it may be necessary to build 2.6.10 natively
from source as I had to. Configure it with

--prefix=/usr/local

to override the package managed version.

GCC v4.6.3 was used for development, at least r5568 is
required due to header file changes. The shared library
system is required.