Added code to the 16-bit streamer so that it transfers to the sound system only as much data has been written into a buffer, rather than the full contents of that buffer which may contain the remnant of a previous buffer-fill operation.
Added speculative code that adds some silence after a streamed file is
closed. It appears on BeagleBoard RISC OS 5.18, a silent tail 3 times
the buffer length is required in order to avoid a nasty buzz if the
file being played does not end in silence. The tail may in fact be
smaller than this, depending on the buffer length or sample rate, but
I found that 2 buffers worth of silence was not enough; 3 was. Iyonix
RISC OS 5.18 exhibits the same problem but to a lesser degree. I
haven't experimented enough to determine a precise formulation that is
sure to work in all cases. And, of course, this code doesn't help when
the DRender: filesystem isn't being used.

Bumped version to 0.56 beta 5.