Addressed a grey area in the documentation concerning which 16-bit channel ordering, left-right or right-left, is "correct" and which is "swapped". The documentation and code now refer explicitly to either left-right or right-left. The default is declared to be right-left.
Disabled the speculative code that adds some silence after a streamed
file is closed. It may have been responsible for a system crash after
the application load caused the desktop to stop multitasking for a
couple of seconds. Besides, adding an arbitrary bit of silence to the
end of a streamed file is a bodgy solution to what I now suspect is an
operating system problem. I discovered that PlayIt suffers from the
same problems on RISC OC 5.18, on BeagleBoard and Iyonix, when using
its acorn16bit driver.

Bumped version to 0.56 beta 6.