Update to V3.0.3.
Additional details:

* mono/utils/riscos-memory.c: Improved memory manager. Use custom memory
management instead of OS_Heap. This allows Mono to allocate some memory,
keep an aligned chunk of it and free the parts before and after that it
doesn't need. With OS_Heap, the memory before and after was wasted.

* mono/metadata/mempool.c: Reverse this commit:

<https://github.com/mono/mono/commit/d900f90f1f69d82f1da4920731c965e44d9b0036>

made to the original source which I believe introduces a bug. The block
of memory allocated should include space for a header and in fact did do,
but this commit removes that space, despite the header still being used,
causing data to overflow. Under RISC OS, this caused the assertion in
mono_class_from_mono_type (mono/metadata/class.c) to fail due to memory
corruption.

* setvars: The class libraries are now built as part of the Mono build.
Install the 4.5 profile.
* mcs/build/config.make: This is now generated by configure.