Update to V3.0.6.
Additional details:

Attempt to improve bootstrapping by building a native version from the
source before the RISC OS patches are applied. This is done using the
preprocess file hook that the Autobuilder provides. A prebuilt mono
package is still required to bootstrap the native build (yes, bootstrap
the bootstrap), but generally this seems to prevent the "corlib out of
sync with runtime" errors.