Update patches to V3.2.1.
Also, patch Makefile.am files instead of Makefile.in. Just to confuse
things, we also have Makefile.am.in in mono/mini and mono/metadata.
I've patched the one in mono/mini as it appears to be copied over
Makefile.am causing the patching to fail, but I didn't see that
occur in the other, so for now I've left that one alone.