Fix data corruption when transferring data to the GPU. This problem was evident on the !Triangle example (the spinning cube) that showed corrupt textures. Tests using an OpenGL version of SDL2 and the game NeverBall also showed corrupt textures and vertices. These particular problems are now fixed. An OpenGL version of Qt5 has also shown significant improvement, but still suffers from a problem where when moving a RISC OS window the rendering origin also moves and the window content disappears (probably a Qt5 specific issue).
* module/main.{c/h}: Create an uncacheable dynamic area to use
as a buffer when transferring data to the GPU. This is divided
into 2 sections. A small section (the merge buffer) for sending
commands and a larger section for transferring bulk data (such
as textures).
* interface/khronos/common/khrn_client.{c/h}: Use uncacheable
memory for the merge buffer.
* interface/vchiq_arm/vchiq_lib.c: Use uncacheable memory for
bulk transfers.

Also:
* interface/khronos/vg/vg_int_mat3x3.c: Comment typo.
* bcm_host/api.c: Version is now 1.02.
* hello_pi/Makefile.include: Build fix.
* inline_swis.h: Move reporter functions...
* reporter.h: ...to here.
* macros.s: For debugging purposes, add the ability to report
function names and arguments.
* module/khronos.cmhg: Increase version number because of the
data corruption fix.
* module/makefile: Extra include path for OSLib.
* wimp_test/main.c: Fix texture distortion when window moved
off left edge of screen by defining the source rectangle when
moving the video core element. This changes the nature of the
example; the triangle no longer changes size to match the
window, so alter the window flags to remove scroll bars and
resize icons.
* !Khronos/Docs/!ReadMe: Minor updates.
* setvars: Change version to 1.02.
* interface/vcos/riscos/vcos_threads.c(ensure_irqs_off,
restore_irqs): Tell the compiler than the flags are corrupted.