Add game Neverball.
This port is targeted at the RaspberryPi and makes use of the
hardware accelerated OpenGL graphics via SDL2/Khronos.

At this time there is no sound support in SDL2 and it's full
screen only.

On the RaspberryPi 3, it seems to average about 40fps.

The packaged/zip'ed size of the game is about 71MB.