GCC 8.2.0 (Experimental)
The intention is to keep this reasonably up to date assuming
it's technically possible (I realise it's already out of date,
but gotta start somewhere!) and keep GCC 4.7.4 as the main compiler
for reasons of compatibility.
A RISC OS native version is on the TODO list.
This version uses a slightly optimised version of the abi-2.0
PIC code. The abi version is "armeabihf".
The name of the compiler is arm-riscos-gnueabihf-gcc or
arm-riscos-gnueabihf-g++.
The compiler does not produce stack checking code, and instead
uses the ARMEABISupport module to provide stack support via
memory mapping/abort handling. There is still some work to be done
on propogating stack overflows to Unixlib.