Update to GCC 10.2.0.
setvars: Be more careful about where we link libdl from. If we point
the linker at GCCSDK_INSTALL_ENV and we don't delete the contents
before rebuilding, there's a danger that we'll pick up a rogue libz
shared library from a past build. Instead point the linker directly at
the libdl we just built in the gcc build tree.
Disable LTO.
Install suitable cross compiler files to be used with Meson and CMake.